Streamline rights comparison
Don't convert arrays to strings just to compare them. Instead compare
them element by element.

Keep the "@id" element in both the metadata and rels-ext
datastreams. Still undecided whether to use `info:fedora/` prefix or not
on the ids. On one hand, that *is* what is stored in fedora. On the
other hand, it is ugly and redundant.