Refactoring OsfToRof for instance ttl_data
Instead of passing the ttl_data as parameters, set an instance
variable. I found it a bit confusing that throughout the code,
references to `ttl_data` could be either the full ttl_data (i.e. an
array) or an element of the original `ttl_data` (i.e. a Hash).

This is my attempt to clean that up and provide greater clarity.