Adding ROF::Translator
I'm looking to normalize the interface of the translators; At present
they have developed organically. As we look to add more into the fold
I want to normalize those interactions -- to remove surprises when
working across translators, keeping layers of behavior separate, and
providing more unified guidance on creating ./bin scripts

Although it is functional, it is a work in progress.