Translating an ETD's JSON-LD to ROF
To orient yourself to the code, review the
`ROF::Translators::JsonldToRof` module.

This is a single tested conversion of one JSON-LD document of an ETD
vs. the Etd node of the ROF document.

Note, this does not yet handle the command line interface of mapping a
JSON-LD document into an ROF document.

DLTP-895

Closes DLTP-906 #Resolve