Fix OSF spec errors
The errors are because the fixture tar file has files with a too-large
uid. The posix tar format only allows for 8 octal digits to represent
the uid for files. Usually users on an ldap system have much higher ids.
This code is using a Rubygems-internal gem to read and write tar files.
The error occurred with this commit that added strict verification of
tar headers:

https://github.com/rubygems/rubygems/commit/92e98bf8f810bd812f919120d4832df51bc25d83

To fix the problem I extracted and recreated the zip file and assigned
each file to have the uid = 0 (root). This is the standard way of
indicating an anonymous user. (This is the "--owner=0" flag in gnu tar).