Redo the json-ld logic
The old algorithm looked at each triple one at a time, and dealt with
the blank nodes as they appeared. This code uses the json/ld gem to
group predicates by subject and then shortens URIs based on our own
prefix table. And triples whose object is a blank node are dealt with
when encountered. Once each triple with the target subject has been
looked at, we map them into the accumulator and use that to generate the
ROF file.

This simplifies the code and makes the parsing and ROF generation two
clearly separate steps.

It also relaxes the error checking and URIs that have prefixes not in
our NAMESPACE map are allowed as predicates (instead of raising an error).
Unknown predicates all end up in the metadata section of the ROF. This
lets us add more predicates as needed without needing to update the code.